[#76125] Testplan.py: Fixed handling of stages in Testpoint/Testplan
The issue was that generated simulation results (and by extension, the
totals for steps) were dependant on the order in which the files were
supplied to the system: if a testplan with one stage appeared first, it
would've included only it's stage in totals, but if it was later in the queue,
where multiple stages were added before it, it would contain totals for
all stages.

This was due to the Testpoint.stages member being a class (essentially,
static) variable and it was globally extended when a new stage was
encountered.

This change handles stages in the context of a single testplan only, as
there was no mechanism that was keeping track of the global state of stages
anyway.

Signed-off-by: Jan Bylicki <jbylicki@antmicro.com>